removing api call tests
Deciding to not implement unit tests for the  Resources. Will be too cumbersome and will not provide much help considering that half the tests would require a valid access token with the permissions, and that the character entity of the access token has all the needed in game assets and state in order to call all the tests.

Will only unit test the client and perhaps the base Resource class.